move integration tests to integration test file
